// src/views/InventoryProduct/utils/type.ts
export interface Product {
  name: string;
  code: string;
}

export interface Inventory {
  id: number;
  product: Product;
  image: string;
  tm: number;
  hn: number;
  th: number;
  sg: number;
  commingProduct: number;
  tn: number;
  vinh: number;
  screen: number;
  thanhhoa: number;
  danang: number;
  xeth: number;
  mrV: number;
}